+++ impl/platform_linux.c
@@ -16,6 +16,7 @@
 #include <pthread.h>
 #include <dlfcn.h>
 #include <errno.h>
+#include <stdlib.h>
 
 void ms_to_ts( u32 ms, struct timespec* out_ts );
 void ms_to_ts_abs( u32 ms, struct timespec* out_ts );
@@ -23,11 +24,19 @@ void ms_to_ts_abs( u32 ms, struct timespec* out_ts );
 void* platform_heap_realloc(
     void* old_buffer, const usize old_size, const usize new_size
 ) {
+    u8* result = (u8*)realloc( old_buffer, new_size );
+    if( !result ) {
+        return NULL;
+    }
+    memory_set( result + old_size, 0, new_size - old_size );
+    return result;
+#if 0
     void* result = mremap( old_buffer, old_size, new_size, MREMAP_MAYMOVE );
     if( result == MAP_FAILED ) {
         return NULL;
     }
     return result;
+#endif
 }
 b32 posix_thread_join_timed( ThreadHandle* handle, u32 ms, int* opt_out_exit_code ) {
     struct timespec ts;